test: add test that checks module interface to expose only allowed fields
Allowed fields are: commands, queries, public events and message bus
