+++ modules/crud/interface.py
@@ -1,5 +1,3 @@
-from dataclasses import dataclass, field
-
 from modules.crud.core.application.commands.messagebox import ProcessInbox
 from modules.crud.core.application.queries.get import GetCrudData
 
@@ -7,9 +5,9 @@ from commons.message_bus.message_bus import MessageBus
 from modules.crud.infrastructure.container import get_container
 
 
-@dataclass
+# @dataclass(frozen=True)
 class Module:
-    message_bus: MessageBus = field(default_factory=lambda: get_container().get(MessageBus), init=False)
+    message_bus = get_container().get(MessageBus)
     GetCrudData = GetCrudData
     ProcessInbox = ProcessInbox
